+++ changes.md
@@ -10,6 +10,10 @@ This document outlines the changes made between versions of the **Goat - SecureM
 
 The automatic update process has been refined with retry logic and better error handling.
 
+#### Web server
+
+Removed unnecessary code which is already provided by the utils module.
+
 ## V1.4.1
 
 ### Bug Fixes